Update tutorials for hosted env (#1)
* Initial updates to support hosted turorials

Change the playground invocation to the call to inject the API
token and setup the tutorials config. Also add an initial paragraph
to bootstrap the environment.

* Add a Taskfile and zip command

Eventually we'll host the zip file here when we make the repo public.
First I will make sure the norebooks are in a working state with the
deployed backend.

* Remove scikit-learn install since it is done later

* Update for the hosted environment backed by snowflake

Add a variable to track the database name since the hardcoded
spark_catalog is no longer an option.

Install featurebyte and pull the prebuilt_catalogs.py script as the
first step.

Also cleanup some artifacts and older notebooks incorrectly committed.