Make the notebooks compatible with the Databricks env
Change to using the %pip line magic command for installing packages.
This ensures packages are installed into the running kernel and is a
necessary change for running in the Databricks notebook env. Also make
sure to install all packages at the start since in Databricks a
package install restarts the kernel and will cause you to lose current
state.